Weird Comparable error
If I have the score compare * -1, then it compares correctly, but
seems to sort wrong?

If I don't have that, it sorts right, but compares incorrectly.

WAIT. Duh.

That's because I'm saying that the Higher the number, the LOWER the priority.

So...DUH.

Okay. Okay okay. I can solve this now. Just got to compare the raw ints